+++ src/it/prova/gestioneproprietari/service/proprietario/ProprietarioServiceImpl.java
@@ -135,7 +135,7 @@ public class ProprietarioServiceImpl implements ProprietarioService {
 	}
 
 	@Override
-	public int contaQuantiProprietariConAutoImmatricolataDopo(int anno) throws Exception {
+	public Long contaQuantiProprietariConAutoImmatricolataDopo(int anno) throws Exception {
 		EntityManager entityManager = EntityManagerUtil.getEntityManager();
 
 		try {
